+++ src/EngineCore/Entities/EntityEventDelegates.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * Sovereign Engine
- * Copyright (c) 2018 opticfluorine
+ * Copyright (c) 2019 opticfluorine
  *
  * Permission is hereby granted, free of charge, to any person obtaining a 
  * copy of this software and associated documentation files (the "Software"), 
@@ -23,38 +23,28 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
-namespace Sovereign.EngineCore.Components
+namespace Sovereign.EngineCore.Entities
 {
 
     /// <summary>
-    /// Defines delegate types for component add/remove/modify events.
+    /// Defines delegate types for entity-related events.
     /// </summary>
-    /// <typeparam name="T">Value type.</typeparam>
-    public static class ComponentEventDelegates<T>
+    public static class EntityEventDelegates
     {
 
         /// <summary>
-        /// Delegate type used to communicate component add and update events.
+        /// Delegate for entity removal events.
         /// </summary>
-        /// <param name="entityId">Entity ID.</param>
-        /// <param name="componentValue">New component value.</param>
-        public delegate void ComponentEventHandler(ulong entityId, T componentValue);
+        /// <param name="entityId">Entity ID that is removed.</param>
+        public delegate void RemoveEntityEvent(ulong entityId);
 
         /// <summary>
-        /// Delegate type used to communicate component remove events.
+        /// Delegate for entity unload events.
         /// </summary>
-        /// <param name="entityId">Entity ID.</param>
-        public delegate void ComponentRemovedEventHandler(ulong entityId);
-
-        /// <summary>
-        /// Delegate type used to communicate component unload events.
-        /// </summary>
-        /// <param name="entityId"></param>
-        public delegate void ComponentUnloadedEventHandler(ulong entityId);
+        /// <param name="entityId">Entity ID that is unloaded.</param>
+        public delegate void UnloadEntityEvent(ulong entityId);
 
     }