+++ src/WorldManagement/Configuration/IWorldManagementConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * Sovereign Engine
- * Copyright (c) 2018 opticfluorine
+ * Copyright (c) 2019 opticfluorine
  *
  * Permission is hereby granted, free of charge, to any person obtaining a 
  * copy of this software and associated documentation files (the "Software"), 
@@ -23,39 +23,23 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
-namespace Sovereign.EngineCore.Components
+namespace Sovereign.WorldManagement.Configuration
 {
 
     /// <summary>
-    /// Defines delegate types for component add/remove/modify events.
+    /// Configuration settings for WorldManagementSystem.
     /// </summary>
-    /// <typeparam name="T">Value type.</typeparam>
-    public static class ComponentEventDelegates<T>
+    public interface IWorldManagementConfiguration
     {
 
         /// <summary>
-        /// Delegate type used to communicate component add and update events.
+        /// Length of a world segment, in blocks.
         /// </summary>
-        /// <param name="entityId">Entity ID.</param>
-        /// <param name="componentValue">New component value.</param>
-        public delegate void ComponentEventHandler(ulong entityId, T componentValue);
-
-        /// <summary>
-        /// Delegate type used to communicate component remove events.
-        /// </summary>
-        /// <param name="entityId">Entity ID.</param>
-        public delegate void ComponentRemovedEventHandler(ulong entityId);
-
-        /// <summary>
-        /// Delegate type used to communicate component unload events.
-        /// </summary>
-        /// <param name="entityId"></param>
-        public delegate void ComponentUnloadedEventHandler(ulong entityId);
+        uint SegmentLength { get; }
 
     }
 
 }
+